requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.0
openai==0.28.1
pypdf==3.15.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.2
